Added an integration test for a static method call to a class that was built using jdk 1.4 or older.  The addition of this test showed that the instrumenting of this type of method was generating incorrect instrumentation code.  The issue was that I was adding the instructions incorrectly.
The above test also uncovered an issue with the Same Local 1 Stack
Item Stack Map Frame issue.  It was incorrectly producing the extended
version when the offset was between 64 and 128.

The integration tests now also check that the spawned process does not
generate an standard error output.  If it does, then the test will fail.