Started the process of splitting up the logger agent into several maven modules.  This is to be able to simplify generating and running unit and integration tests.
This set of changes includes the separating of the Logger classes
from the agent instrumentation classes.  The classes that are being
injected into the instrumented classes will now reside in the
plogger-agent-logger module.  This also includes this module specific
unit tests.  As part of this work, more unit tests were added that
tests other features of those classes.

Moved some of the integration tests from the plogger-agent module
into the plogger-agent-integration-tests module.  These tests are
currently not be used.  There are other changes that require them
to be able to run as expected.  This work will be done in
subsequent commits.