The command line menu command will print invalid menu item selection to standard error instead of standard out.  This is to allow the testing to be able to determine if an invalid menu item was entered.
Updated the tests mechanism to be able to determine if an
invalid menu selection was entered.

The SendOptions class had a bug in which it was not
resetting the _option field to ''.  This was causing
invalid tests to pass.  This has been corrected.  Added
ability to check if a given output was generated by the
called menu items.  This allows us to be able to determine
if the correct menu item was called within a given test.

Finally, added a new test that will called a second and
third level menu.  It finally calls a menu item that is
not a sub menu and determine if the expected method was
called by checking the generated output from the menu
calls.