The following changes contains the main work to be able to get the plugin update a list with the different times that has been spent on a given web page.
The current incarnation is able to correctly update a simple
unordered list with this information.

There is more work that needs to be implemented that is related to
the opened window.  It is currently a very simple web page that is
not user friendly.

The other work that is required is that this page should be opened
by the users of the plugin and not opened when the plugin is
initially loaded.

I'm sure that there are other work involved but at this point this
work is a good proof of concept.