Reverted the original css settings used by beautiful jekyll to be consistent with what was initially used by this web site.
Created a separate resume.css file that defines the table formatting
values and only include this as part of the pages that require the
format.  The other pages that don't will just default to the settings
of what was set for the beautiful jekyll theme.

Added the tunn-on-fill-column command to all of the files that require
this to be done automatically.  This removes the need to form to use
the command every time.

Continued to update several of the resume specific files with my
experience of the different tools, concepts, languages and whatnot.

There is also the change that replaced the default formatting of the
page to reflect one that uses more of the windows space for the area
of the window that is used by my writting.