Added more implementation of search algorithms that include the uniform cost search algorithm.  The addition of this algorithm revealed that the current use of the node class isn't such a good idea.  This in turn will require that all the search algorithms need to be change to allow for a better implementation that code also be used by the uniform cost search algorithm.  It might be the case that that algorithm and others similar to that would need a different node/edge implementation than the other search algorithms.
Added more tests to the operators project that is used to learn and test
different operator implementations.

Added a shared library that will be used by the other projects.  This
initiallly includes some debugging statements and generate macros that
can be used by the other projects.

Lastly, I have added a README file to the search project that will
discuss the different implementations for the search project.