Created a different version of a depth first search that was implemented within the breath first search source code.  It is a first step towards trying to create something that would looking at the available memory and determine if it needs to store some of the results on disk before continuing to perform the search.
As part of that check, a resource type classes are being added that will
be used to determine how much memory the process and/or system is using
so that the search will take appropiate actions that will allow it to
continue to perform the search.

These added checks will only work on windows so at the moment the other
operating system tests will fail.

Note that none of the above changes are complete solutions but only a
step towards that objective.  Thus, these changes will cause test builds
and errors.