Added the problem solutions to the following techie delight problems:
	- Largest Consecutive Subarray
	- Longest Bitonic Subarray
	- Majority Element
	- Move Zeroes to End
	- Replace Array Elements

These solutions include different tests that were used to insure that
the solution was correct.  There are tests that were added because
those are the ones that failed when they were submitted.

Created a template check that is used for checking that a solution is
one of a list of solutions.  This implementation started to be used by
more than one of the solution test implementations.

While there are other minors changes that reflected the updating of
version numbers and ignoring of the Testing directories.