Added the following implementations that are not solutions to the following problems:
	- Interleaving String
	- Interleaving String II

As part of the above additions, a new way was added to include different
combinations of input and outputs.  It is now possible to create a data
file that will contain each test input and output values.  This removes
the need to have to recompile the test sources whenever a new
input/output test is being added.  This also makes it such that the only
source file that will be altered is the implementation and not its test
source file.

The Loader class was added that is used to read the different input and
output data that will be used by the different tests.  This class
requires that the importing class will have its >> operator implemented
and that the class has the no parameter constructor defined.

There is also the find_path method that was added to be able to find the
data file.  This was added because of the fact that the tests don't have
to be run within specific directory but that it just needs to be
executed within any of the project directory.