Added solutions to the problems:
	- N Queen Problem
	- Knight Tour Problem

While the solutions to the problems above are correct.  One of the tests
for the knight tour problem is failing even though it generated the
correct answer.  Apparently, the ASSERT_EQ macro using a set instances
doesn't work as expected.  Using the ASSERT_THAT macro and the
AllOfArray matcher didn't compile.  While the creation of a user defined
matcher class, ContainersAll also failed to correctly compare the actual
and expected values.  Thus this particular test will fail until a
solution to this test issue is resolved.

There is also another point that needs to be noted.  The solution to the
knight tour problem works but feels that it is too slow.  Thou the
solution does pass the conditions for the implementation on techie
delight.  It doesn't feel like this inplementation is not that much
slower than a most performant version of the solution.   I will need to
compare a performant version with this version.