Added solutions to the following problems:
	- K Sum Subarray
	- K Sum Subarray II
	- Smallest Subarray
	- Smallest Window to Sort Array

While the first three have complete solutions.  The solution to Smallest
Window to Sort Array's solution is not totally correct.  This will
require another go to get the correct solution.

As part of the above solutions, the pairLoader template class was
created that is used to load pair instances from an istream.

Lastly, the README file was updated for the above implementations.