Added solutions to the following problems:
	- Buy N Sell Shares
	- Maximum Path Sum
	- Minimum Platforms

But not all of the above are complete solution.  The Maximum Path Sum
solution doesn't not solve all cases and needs work to solve these
remaining cases.  Those cases are part of the tests that have been added
to that solution tests.

Finally, the following solution was not implemented because I don't have
an idea on how to solve this problem.

	- Trapping Rain Water

Hopefully, a solution will come to me.

The above information was added to the README file.