Added solutions to the following problems:
	- Sort Array Single Swap
	- 3 Sum
	- 3 Sum II
	- 3 Sum III

While the above solutions are correct and are also fast.  I feel that
the different solutions to the 3 Sum problems can be improved such that
a non recursive solution can be used.  This is something that will be
done in another version of these solutions.

Update the README file to reflect these solutions.