Added solutions to the following problems:
	- Geometric Triplets
	- Buy and Sell Stock

But the Buy and Sell Stock is partially implemented.  It does need some
more work to be able to handle other test cases.

With the above solutions, their were changes made to the checker and
debug classes.  This included adding a skipSpace struct that will skip
all space like characters.  This was added to make its use easier.  The
changes to the debug struct is to be able to test that the
implementation works as expected.  As part of those changes, a test was
dded that checks that the entries are being logged properly.