Added solutions to the problems:
	- Buy and Sell Stock II
	- Buy and Sell Stock III

with the caveat that the Buy and Sell Stock II is a partial solution to
that problem.

The above information is reflected in the README file.

The Buy and Sell Stock III solution is correct but it might be
reimplemented in such a way that it doesn't require a recursive
implementation.