Added solutions to the following problems:
	- Frequency Map Two
	- Shuffle Array
	- Shuffle Array II
	- Shuffle Array III
	- Sorted Triplet

All of the above solutions are complete except for the Shuffer Array III
solution.  This is only a partial solution.

The above solutions have been noted in the README file.

As part of the work above, a new triple tuple loader was introduced that
was used to load data.  It is hoped in the future that this
implementation can be extended to be able to load any sized tuple.