Cleaned up the loader template code such that it will allow the user to be able to use a more compact data definition.  The original implementation would read to the end of line.  This was limiting the user such that they had to define the next set of data on a new line. This requirement has been removed for the different loaders.
As part of this work, the original vectorLoader and matrixLoader
implementation has be set to deprecated.  It is the intention to replace
all references of these class with the newer version.  Upon all
replacements have been done.  These classes will then be removed from
the distribution to insure that the newer version will be used instead.