Added solutions to the problems:
	- Coin Change Problem
	- Coin Change Problem II
	- Minimum Merge Operations
	- Next Greater Element
	- Next Greater Element II
	- Search Nearly Sorted Array

Note that the last solution improves on its prior solution that was not
correctly producing the expected results.  This version handles the
boundary cases that the prior solution was not able to resolve.

Lastly, fixed a compiler error in which the three partition problem
include file prefixed the class method with the class name when that is
not supposed to be included within the definition of the class in the
include file.

All of the above has been reflected in the README file.