Added the following problem implementations:
	- Rod Cutting
	- Rod Cutting II
	- Unattended Customers Count
	- Minimum Number K Swaps
	- Shrink Array

Out of the five a solution is provided for only the Rod Cutting II and
Shrink Array problems.  While there is a partial solution to the Minimum
Number K Swaps and Unattended Customers Count problem.  Lastly the Rod
Cutting problem doesn't have an implementation.

This is all reflected in the README file.