Added a new std::tuple loader that will be able to create a n-tuple instead of the 3-nuple loader.  This new implementation is a more general implementation that one can use when loading different types of entries for the defined tuple.  As part of this addition, tests have been created that tests the new functionality.
The above implementation inspired the creation a general loader class,
dataLoader.  This work led to multiple version of the dataLoader.  At
the moment, there are 4 versions that have been implemented.  They are
all included here for historical reasons at the moment but in the future
the best version will remain and the others will be removed.  They all
are part of the valhalla::utils::loaders::loader::v[1-4] namespaces.
Each of the different template classes require different template
parameters.  The specifics can be surmised by looking the data_loader.h
file or the dataloadertests.cpp files.  As part of the addition of the
different dataLoader implementations, tests have been defined that test
different configurations that the dataLoader implementation is supposed
to be able to handle.