Added skeleton solutions to the problems:
	- Balanced Expression
	- Group Anagrams
	- Transform String
	- Transform String II
	- Transform String III

As part of the above, tests were added to insure that these solutions
will be looked at in the future.

The above changes are reflected in the README file.

As part of the above skeleton solutions.  A new utility template was
added, is_space_noop, that is used to not skip any characters when
looking for space like characters to skip.  This was added to allow
for reading a string that is surrounded by '"' character or the likes.