Added a version 6 of the dataLoader template class that uses the if constexpr preprocessing C++ feature to only generate the require code that is needed to process the input data.  This will produce more compact code whenever we are loading primitive types compared to complex types.  This will thus improve the performance of the dataLoader by removing the need to generate extra layer of code to be called whenever we are processing primitive data.
Enhanced the skip_spaces template struct to return a reference to
the passed basic_istream<char_type> data to make consistent with the
expected implementation of the input operator <<.  This will allow
the code to chain these calls together and make them more consistent
with the input operator << protocol.

As part of the above changes, several tests have been added to check
that the current implementation will work like the prior versions of
the dataLoader template.  As part of these tests, new ones were
included that will test other features that were not tested using
the prior versions of the dataLoader template class.