This commit contains the initial commit of the dataloader template class was part of the cpp_data_structs_and_algo branch to its own branch. This is to allow others to benefit from this work.
As part of this work, the dataloader template class is using a different
namespace than what was part of the original version.  This also
includes updates to the tests themselves.

While most of the code is similar to the original version there were
some changes are not part of the original version.  This includes the
addition of the is_space_base template class that replaces the need to
build a library that would be used with the template classes.

This commit completes the work for issue #2.