This set of changes contains updates to my work experience and self-learning that I have done that could be beneficial to someone that would be interested in hiring me.
This is also the addition of the notes section that contains different
notes that I am trying to put together that could be useful for me and
others in the future.  While this is a first pass, it is possible that
the format might change in the future.  That is not all since this work
in still "in progress" and more changes will be added.

Lastly, I added some more quotes and updated the quotes javascript to
display the author information to the left and right of the quote.  This
also includes a preliminary go at a remote work post that needs to be
completed prior to being published.