Added the string example that shows how you can use the data loader template class to be able to load a string that is surrounded by a " character and also how to process an escape character.
As part of the above work, the README files for the examples have been
extended to include more details about each example that is part of the
distribution.

The above also included some updates to the examples.  These changes
mostly involved the addition or rewording of the source code comments to
provide more detail about each example.