Added a failing container test for v6 of the dataloader.  This test will test if the dataloader can load an empty container in which the opening and closing characters contain spaces between them.
As part of the above work, it was noted that the BasicStringReader
should be performing a while loop instead of an if statement.  This was
corrected.

All of the dataloader tests pass except for the newly added one which is
expected at the moment.  This test was added such that we don't forget
to determine a fix for this type of issue.