Implemented a solution for the following problem:
   - N Digit Numbers VII

The solutions include more tests that stresses the implementation.

The above information was added to the README file.

As part of the above, the following solution to the following problem
was left out recently.

   - Wildcard Pattern Matching II

The above solution was mention in the README but the code was not
included as part of the commit.