Implemented a solution for the following problems:
   - Add Binary
   - Circular Shift
   - XOR Without XORing
   - Power Set
   - Power Set II

The solutions include more tests that stresses the implementations.

The above information was added to the README file.

The above also includes an implementation that generates n instances
of data that will be used to test a given feature.  The
circular_shifttests.cpp file contains this implementation that needs to
be extracted and placed within its own source file.  The moved code
should be generic, i.e., templated.